+++ test/template_test.exs
@@ -29,6 +29,48 @@ defmodule SparkPost.TemplateTest do
     end
   end
 
+  describe "Template.list" do
+    test_with_mock "returns list of templates", HTTPoison,
+      request: fn method, url, body, headers, opts ->
+        assert method == :get
+        assert url =~ "/templates"
+        fun = MockServer.mk_http_resp(200, MockServer.get_json("listtemplate"))
+        fun.(method, url, body, headers, opts)
+      end do
+      assert [%SparkPost.Template{}, %SparkPost.Template{}] = Template.list()
+    end
+
+    test_with_mock "passes :draft param", HTTPoison,
+      request: fn method, url, body, headers, opts ->
+        assert opts[:params] == [draft: true]
+        fun = MockServer.mk_http_resp(200, MockServer.get_json("listtemplate"))
+        fun.(method, url, body, headers, opts)
+      end do
+      assert Template.list(%{draft: true})
+    end
+
+    test_with_mock "passes :shared_with_subaccounts param", HTTPoison,
+      request: fn method, url, body, headers, opts ->
+        assert opts[:params] == [shared_with_subaccounts: false]
+        fun = MockServer.mk_http_resp(200, MockServer.get_json("listtemplate"))
+        fun.(method, url, body, headers, opts)
+      end do
+      assert Template.list(%{shared_with_subaccounts: false})
+    end
+
+    test_with_mock "returns error response", HTTPoison,
+      request: fn method, url, body, headers, opts ->
+        fun = MockServer.mk_error("Uknown")
+        fun.(method, url, body, headers, opts)
+      end do
+      assert Template.list() == %SparkPost.Endpoint.Error{
+               status_code: nil,
+               errors: ["Uknown"],
+               results: nil
+             }
+    end
+  end
+
   test_with_mock "Template.preview succeeds with Content.Inline",
                  HTTPoison,
                  request: fn method, url, body, headers, opts ->